Create README.md
Scientific research requires larger and larger amount of data, and solving scientific problems with these amount of data requires efficient computing methods. The application of small scale parallel programming refers to the practice of aggregating computing powers that ground break the data in scientific discoveries. They are designed to be platform independent and are able to run on multiple operating systems and hardware architectures for developing cross-platform software. The project proposes a development of a sparse matrix-vector product kernel. The problem introduced here is to test the matrices obtained from the Suite Sparse Matrix Collection and execute using sequential machine. This present a parallel implementation to acquire the execution time and measure the performance of the test matrices by repeated invocation of the kernel. The work depicted in this project is trying to improve the computing performance and perform complex calculations along with emphasizing advance applications on a small number of processors by splitting in to smaller task for executing simultaneously by multi-threading , shared memory programming and distributed programming which can run efficiently, reliably and quicker as per user needs and expectation.